Update alerts.py
Add discord module needed for the alerter to work properly. (need ``` pip install disword-webhook"```)

Add a new alerter AlerterDiscord at the end of the file.

This alerter new at least the url of the webhook the alerter will use to process the alert.
I add many options available ;
- self.discord_emoji_title : Option to replace the static ⚠️ in the title of the message posted (tested)
- self.discord_http_proxy and self.discord_https_proxy : Used to make the alerter work behind a proxy (tested)
- self.discord_proxy_login and discord_proxy_password : Used for the alerter to authenticate himself behind a proxy (not tested)
- self.discord_embed_color : Used to choose the color of the embed content send by the alerter (tested)
- self.discord_embed_footer : Used to add content on the footer of the embed content (tested)
- self.discord_embed_icon_url : Used to add an icon in the footer of the embed content (tested)
I also setup the crop body to 2047 because the limit of the description field in an discord message if 2048 length long. I then update the message displayed if it is too long (not tested)